GameCode
Our task was to play this lander game and collect data while safely landing the lander. Based on the data collected we designed, trained feedforward and backpropagation neural network and based on the test dataset you optimized the hyper parameters. Output of the this project was that lander will automatically land at the right place and safely.